refactor(ui): extract project profile role modals
