+++ docs-ai/skills/ui/SKILL.md
@@ -175,8 +175,10 @@ Each section has exactly one job: orient, inspect, compare, edit, or confirm. If
   `useProjectAssistantController`; do not put propose/apply/bootstrap request
   state back into `ProjectsView`. The workspace rendering branches live behind
   `ProjectWorkspaceView`, and project defaults/root editing lives in
-  `ProjectSettingsPanel` plus `CreateProjectWorktreeModal`, while project state
-  loading and mutations stay in the parent page.
+  `ProjectSettingsPanel` plus `CreateProjectWorktreeModal`. Agent profile and
+  project role editing lives in `ProfilesModal` and `RolesModal`, with shared
+  profile/role form mapping in `projectProfilesRoles.ts`; keep project state
+  loading and mutations in the parent page.
 - External Agent readiness belongs in Connections and in the picker
   diagnostics: distinguish missing binaries, auth/billing problems, unsupported
   versions, and managed-launcher issues without sending users to raw logs first.
