+++ ui/src/features/projects/ProjectsView.tsx
@@ -79,6 +79,8 @@ import {
 } from "./projectAssignmentViewModels";
 import { ProjectAssistantPanel } from "./ProjectAssistantPanel";
 import { CreateProjectWorktreeModal } from "./CreateProjectWorktreeModal";
+import { ProfilesModal } from "./ProfilesModal";
+import { RolesModal } from "./RolesModal";
 import { useProjectAssistantController } from "./useProjectAssistantController";
 import type {
   ProjectAssignmentRecord,
@@ -101,11 +103,7 @@ import type {
   UpdateProjectSkillPayload,
   UpdateProjectWorkItemPayload,
 } from "../../types/project";
-import type {
-  AgentProfileRecord,
-  CreateAgentProfilePayload,
-  UpdateAgentProfilePayload,
-} from "../../types/agent-profile";
+import type { AgentProfileRecord } from "../../types/agent-profile";
 import type { ContextPacketRecord } from "../../types/context";
 import {
   Badge,
@@ -123,6 +121,15 @@ import {
   type CreateWorktreeForm,
   type ProjectDefaultsForm,
 } from "./projectSettings";
+import {
+  profileCreatePayloadFromForm,
+  profileUpdatePayloadFromForm,
+  projectSkillStatusRank,
+  rolePayloadFromForm,
+  splitIDs,
+  type AgentProfileForm,
+  type RoleForm,
+} from "./projectProfilesRoles";
 
 type Props = {
   onOpenChat?: (request: ProjectAssignmentChatLaunchRequest) => void;
@@ -247,37 +254,6 @@ type SkillForm = {
   trustLabel: string;
 };
 
-type AgentProfileForm = {
-  id: string;
-  name: string;
-  description: string;
-  instructions: string;
-  surface: string;
-  providerHint: string;
-  modelHint: string;
-  executionProfile: string;
-  toolsEnabled: boolean;
-  writesAllowed: boolean;
-  networkAllowed: boolean;
-  approvalPolicy: string;
-  projectMemoryPolicy: string;
-  contextSourcePolicy: string;
-  skillIDs: string;
-  externalAgentKind: string;
-};
-
-type RoleForm = {
-  id: string;
-  name: string;
-  description: string;
-  instructions: string;
-  defaultDriverKind: string;
-  defaultProvider: string;
-  defaultModel: string;
-  defaultAgentProfile: string;
-  skillIDs: string;
-};
-
 const WORK_ITEM_STATUSES = [
   "backlog",
   "ready",
@@ -304,10 +280,6 @@ const MEMORY_TRUST_LABELS = [
   "external_untrusted",
   "runtime_state",
 ];
-const AGENT_PROFILE_SURFACES = ["any", "hecate_chat", "hecate_task", "external_agent"];
-const AGENT_PROFILE_APPROVAL_POLICIES = ["inherit", "require", "block", "allow"];
-const AGENT_PROFILE_MEMORY_POLICIES = ["inherit", "include", "visible_only", "exclude"];
-const AGENT_PROFILE_CONTEXT_POLICIES = ["inherit", "include_enabled", "visible_only", "exclude"];
 const MEMORY_SOURCE_KINDS = [
   "operator",
   "generated",
@@ -4227,732 +4199,6 @@ function ProjectRootSelect({
   );
 }
 
-function ProfilesModal({
-  error,
-  pending,
-  profiles,
-  project,
-  projectSkills,
-  roles,
-  onClose,
-  onCreate,
-  onDelete,
-  onUpdate,
-}: {
-  error: string;
-  pending: boolean;
-  profiles: AgentProfileRecord[];
-  project: ProjectRecord;
-  projectSkills: ProjectSkillRecord[];
-  roles: ProjectWorkRoleRecord[];
-  onClose: () => void;
-  onCreate: (
-    form: AgentProfileForm,
-  ) => AgentProfileRecord | undefined | Promise<AgentProfileRecord | undefined>;
-  onDelete: (profile: AgentProfileRecord) => boolean | Promise<boolean>;
-  onUpdate: (
-    profileID: string,
-    form: AgentProfileForm,
-  ) => AgentProfileRecord | undefined | Promise<AgentProfileRecord | undefined>;
-}) {
-  const [selectedProfileID, setSelectedProfileID] = useState(profiles[0]?.id ?? "new");
-  const selectedProfile = profiles.find((profile) => profile.id === selectedProfileID) ?? null;
-  const editingNew = selectedProfileID === "new";
-  const [deleteProfile, setDeleteProfile] = useState<AgentProfileRecord | null>(null);
-  const [form, setForm] = useState<AgentProfileForm>(() =>
-    selectedProfile ? profileFormFromRecord(selectedProfile) : emptyAgentProfileForm(),
-  );
-
-  function selectProfile(profileID: string) {
-    setSelectedProfileID(profileID);
-    const profile = profiles.find((item) => item.id === profileID) ?? null;
-    setForm(profile ? profileFormFromRecord(profile) : emptyAgentProfileForm());
-  }
-
-  function selectProfileRecord(profile: AgentProfileRecord) {
-    setSelectedProfileID(profile.id);
-    setForm(profileFormFromRecord(profile));
-  }
-
-  const canSave = form.name.trim().length > 0;
-  const submit = async () => {
-    if (!canSave) return;
-    if (editingNew) {
-      const profile = await onCreate(form);
-      if (profile) selectProfileRecord(profile);
-      return;
-    }
-    const profile = await onUpdate(selectedProfileID, form);
-    if (profile) selectProfileRecord(profile);
-  };
-
-  async function deleteSelectedProfile(profile: AgentProfileRecord) {
-    const deleted = await onDelete(profile);
-    if (!deleted) return;
-    setDeleteProfile(null);
-    const nextProfile = profiles.find((item) => item.id !== profile.id) ?? null;
-    if (nextProfile) {
-      selectProfileRecord(nextProfile);
-      return;
-    }
-    setSelectedProfileID("new");
-    setForm(emptyAgentProfileForm());
-  }
-
-  return (
-    <>
-      <Modal
-        title="Agent profiles"
-        onClose={onClose}
-        width={840}
-        footer={
-          <div style={{ display: "flex", gap: 8, width: "100%" }}>
-            {selectedProfile && !editingNew && (
-              <button
-                className="btn btn-ghost"
-                type="button"
-                disabled={pending}
-                onClick={() => setDeleteProfile(selectedProfile)}
-                style={{ color: "var(--red)" }}
-              >
-                Delete profile
-              </button>
-            )}
-            <button
-              className="btn btn-primary"
-              type="button"
-              disabled={pending || !canSave}
-              onClick={() => void submit()}
-              style={{ marginLeft: "auto" }}
-            >
-              {pending ? "Saving..." : editingNew ? "Create profile" : "Save profile"}
-            </button>
-          </div>
-        }
-      >
-        <div style={{ display: "grid", gridTemplateColumns: "220px 1fr", gap: 14, minHeight: 470 }}>
-          <div
-            style={{
-              borderRight: "1px solid var(--border)",
-              paddingRight: 10,
-              display: "grid",
-              alignContent: "start",
-              gap: 6,
-            }}
-          >
-            <button
-              className={
-                selectedProfileID === "new" ? "btn btn-primary btn-sm" : "btn btn-ghost btn-sm"
-              }
-              type="button"
-              onClick={() => selectProfile("new")}
-              style={{ justifyContent: "flex-start" }}
-            >
-              <Icon d={Icons.plus} size={12} />
-              New profile
-            </button>
-            {profiles.map((profile) => (
-              <button
-                key={profile.id}
-                className={
-                  selectedProfileID === profile.id
-                    ? "btn btn-primary btn-sm"
-                    : "btn btn-ghost btn-sm"
-                }
-                type="button"
-                onClick={() => selectProfile(profile.id)}
-                style={{ justifyContent: "flex-start", minWidth: 0 }}
-              >
-                <span style={{ overflow: "hidden", textOverflow: "ellipsis" }}>
-                  {profile.name || profile.id}
-                </span>
-              </button>
-            ))}
-          </div>
-          <form
-            onSubmit={(event) => {
-              event.preventDefault();
-              void submit();
-            }}
-            style={{ display: "grid", gap: 12, alignContent: "start" }}
-          >
-            {error && <InlineError message={error} />}
-            <div style={{ display: "grid", gridTemplateColumns: "160px 1fr", gap: 10 }}>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Profile id</span>
-                <input
-                  className="input"
-                  value={form.id}
-                  disabled={!editingNew}
-                  placeholder="implementation"
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, id: event.target.value }))
-                  }
-                />
-              </label>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Name</span>
-                <input
-                  className="input"
-                  value={form.name}
-                  autoFocus={editingNew}
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, name: event.target.value }))
-                  }
-                />
-              </label>
-            </div>
-            <label style={fieldStyle}>
-              <span style={fieldLabelStyle}>Description</span>
-              <textarea
-                className="input"
-                value={form.description}
-                rows={2}
-                onChange={(event) =>
-                  setForm((current) => ({ ...current, description: event.target.value }))
-                }
-              />
-            </label>
-            <label style={fieldStyle}>
-              <span style={fieldLabelStyle}>Instructions</span>
-              <textarea
-                className="input"
-                value={form.instructions}
-                rows={5}
-                onChange={(event) =>
-                  setForm((current) => ({ ...current, instructions: event.target.value }))
-                }
-              />
-            </label>
-            <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 10 }}>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Surface</span>
-                <select
-                  className="input"
-                  value={form.surface}
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, surface: event.target.value }))
-                  }
-                >
-                  {AGENT_PROFILE_SURFACES.map((surface) => (
-                    <option key={surface} value={surface}>
-                      {surface}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Execution profile</span>
-                <input
-                  className="input"
-                  value={form.executionProfile}
-                  placeholder="implementation"
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, executionProfile: event.target.value }))
-                  }
-                />
-              </label>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Provider hint</span>
-                <input
-                  className="input"
-                  value={form.providerHint}
-                  placeholder="ollama"
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, providerHint: event.target.value }))
-                  }
-                />
-              </label>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Model hint</span>
-                <input
-                  className="input"
-                  value={form.modelHint}
-                  placeholder="qwen2.5-coder"
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, modelHint: event.target.value }))
-                  }
-                />
-              </label>
-            </div>
-            <div style={{ display: "flex", gap: 12, flexWrap: "wrap" }}>
-              <label style={checkboxLabelStyle}>
-                <input
-                  type="checkbox"
-                  checked={form.toolsEnabled}
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, toolsEnabled: event.target.checked }))
-                  }
-                />
-                Tools enabled
-              </label>
-              <label style={checkboxLabelStyle}>
-                <input
-                  type="checkbox"
-                  checked={form.writesAllowed}
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, writesAllowed: event.target.checked }))
-                  }
-                />
-                Writes allowed
-              </label>
-              <label style={checkboxLabelStyle}>
-                <input
-                  type="checkbox"
-                  checked={form.networkAllowed}
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, networkAllowed: event.target.checked }))
-                  }
-                />
-                Network allowed
-              </label>
-            </div>
-            <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 10 }}>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Approval policy</span>
-                <select
-                  className="input"
-                  value={form.approvalPolicy}
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, approvalPolicy: event.target.value }))
-                  }
-                >
-                  {AGENT_PROFILE_APPROVAL_POLICIES.map((policy) => (
-                    <option key={policy} value={policy}>
-                      {policy}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Memory policy</span>
-                <select
-                  className="input"
-                  value={form.projectMemoryPolicy}
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, projectMemoryPolicy: event.target.value }))
-                  }
-                >
-                  {AGENT_PROFILE_MEMORY_POLICIES.map((policy) => (
-                    <option key={policy} value={policy}>
-                      {policy}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>Context source policy</span>
-                <select
-                  className="input"
-                  value={form.contextSourcePolicy}
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, contextSourcePolicy: event.target.value }))
-                  }
-                >
-                  {AGENT_PROFILE_CONTEXT_POLICIES.map((policy) => (
-                    <option key={policy} value={policy}>
-                      {policy}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <label style={fieldStyle}>
-                <span style={fieldLabelStyle}>External agent kind</span>
-                <input
-                  className="input"
-                  value={form.externalAgentKind}
-                  placeholder="claude_code"
-                  onChange={(event) =>
-                    setForm((current) => ({ ...current, externalAgentKind: event.target.value }))
-                  }
-                />
-              </label>
-            </div>
-            <SkillIDPicker
-              onChange={(skillIDs) => setForm((current) => ({ ...current, skillIDs }))}
-              skills={projectSkills}
-              value={form.skillIDs}
-            />
-            <div style={subtleTextStyle}>
-              Profiles set runtime posture and skill references. Skills do not grant tools, writes,
-              network, or approvals.
-            </div>
-          </form>
-        </div>
-      </Modal>
-      {deleteProfile && (
-        <ConfirmModal
-          title="Delete agent profile"
-          danger
-          pending={pending}
-          confirmLabel="Delete agent profile"
-          onClose={() => setDeleteProfile(null)}
-          onConfirm={() => void deleteSelectedProfile(deleteProfile)}
-          message={
-            <>
-              Delete <strong>{deleteProfile.name || deleteProfile.id}</strong>.{" "}
-              {profileReferenceSummary(deleteProfile, project, roles)} Other projects may also
-              reference this global profile.
-            </>
-          }
-        />
-      )}
-    </>
-  );
-}
-
-function SkillIDPicker({
-  disabled = false,
-  skills,
-  value,
-  onChange,
-}: {
-  disabled?: boolean;
-  skills: ProjectSkillRecord[];
-  value: string;
-  onChange: (value: string) => void;
-}) {
-  const selectedIDs = uniqueSkillIDs(splitIDs(value));
-  const selectedSet = new Set(selectedIDs);
-  const indexedSkills = new Map(skills.map((skill) => [skill.id, skill]));
-  const sortedSkills = sortProjectSkillsForPicker(skills);
-  const warnings = selectedIDs.flatMap((id) => projectSkillSelectionWarnings(id, indexedSkills));
-
-  function toggleSkill(skillID: string, checked: boolean) {
-    const next = checked
-      ? uniqueSkillIDs([...selectedIDs, skillID])
-      : selectedIDs.filter((id) => id !== skillID);
-    onChange(next.join(", "));
-  }
-
-  return (
-    <div style={fieldStyle}>
-      {sortedSkills.length > 0 && (
-        <div style={{ display: "grid", gap: 6 }}>
-          <span style={fieldLabelStyle}>Project skills</span>
-          <div style={{ display: "grid", gap: 6 }}>
-            {sortedSkills.map((skill) => (
-              <label
-                key={`${skill.id}:${skill.path}`}
-                style={{
-                  border: "1px solid var(--border)",
-                  borderRadius: 6,
-                  display: "grid",
-                  gap: 4,
-                  gridTemplateColumns: "auto 1fr",
-                  padding: "7px 8px",
-                }}
-              >
-                <input
-                  type="checkbox"
-                  checked={selectedSet.has(skill.id)}
-                  disabled={disabled}
-                  aria-label={`Use skill ${skill.title || skill.id}`}
-                  onChange={(event) => toggleSkill(skill.id, event.target.checked)}
-                  style={{ marginTop: 2 }}
-                />
-                <span style={{ display: "grid", gap: 4, minWidth: 0 }}>
-                  <span style={{ display: "flex", gap: 6, alignItems: "center", flexWrap: "wrap" }}>
-                    <span style={titleStyle}>{skill.title || skill.id}</span>
-                    <span className={projectSkillBadgeClass(skill)}>{skill.status}</span>
-                    {!skill.enabled && <span className="badge badge-muted">disabled</span>}
-                    <span className="badge badge-muted">{skill.id}</span>
-                  </span>
-                  <span style={subtleTextStyle}>{skill.path}</span>
-                </span>
-              </label>
-            ))}
-          </div>
-        </div>
-      )}
-      <label style={{ ...fieldStyle, marginTop: sortedSkills.length > 0 ? 8 : 0 }}>
-        <span style={fieldLabelStyle}>Skill ids</span>
-        <input
-          className="input"
-          value={value}
-          disabled={disabled}
-          placeholder="backend, qa"
-          onChange={(event) => onChange(event.target.value)}
-        />
-      </label>
-      {warnings.length > 0 && (
-        <div style={{ display: "grid", gap: 3 }}>
-          {warnings.map((warning) => (
-            <div key={warning} style={{ ...subtleTextStyle, color: "var(--amber)" }}>
-              {warning}
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function RolesModal({
-  agentProfiles,
-  error,
-  pending,
-  projectSkills,
-  roles,
-  onClose,
-  onCreate,
-  onDelete,
-  onUpdate,
-}: {
-  agentProfiles: AgentProfileRecord[];
-  error: string;
-  pending: boolean;
-  projectSkills: ProjectSkillRecord[];
-  roles: ProjectWorkRoleRecord[];
-  onClose: () => void;
-  onCreate: (
-    form: RoleForm,
-  ) => ProjectWorkRoleRecord | undefined | Promise<ProjectWorkRoleRecord | undefined>;
-  onDelete: (role: ProjectWorkRoleRecord) => boolean | Promise<boolean>;
-  onUpdate: (
-    roleID: string,
-    form: RoleForm,
-  ) => ProjectWorkRoleRecord | undefined | Promise<ProjectWorkRoleRecord | undefined>;
-}) {
-  const customRoles = roles.filter((role) => !role.built_in);
-  const firstEditable = customRoles[0] ?? null;
-  const [selectedRoleID, setSelectedRoleID] = useState(firstEditable?.id ?? "new");
-  const selectedRole = roles.find((role) => role.id === selectedRoleID) ?? null;
-  const editingBuiltIn = Boolean(selectedRole?.built_in);
-  const editingNew = selectedRoleID === "new";
-  const [form, setForm] = useState<RoleForm>(() =>
-    selectedRole ? roleFormFromRecord(selectedRole) : emptyRoleForm(),
-  );
-
-  function selectRole(roleID: string) {
-    setSelectedRoleID(roleID);
-    const role = roles.find((item) => item.id === roleID) ?? null;
-    setForm(role ? roleFormFromRecord(role) : emptyRoleForm());
-  }
-
-  function selectRoleRecord(role: ProjectWorkRoleRecord) {
-    setSelectedRoleID(role.id);
-    setForm(roleFormFromRecord(role));
-  }
-
-  const canSave = form.name.trim().length > 0 && !editingBuiltIn;
-  const submit = async () => {
-    if (!canSave) return;
-    if (editingNew) {
-      const role = await onCreate(form);
-      if (role) {
-        selectRoleRecord(role);
-      }
-      return;
-    }
-    const role = await onUpdate(selectedRoleID, form);
-    if (role) {
-      selectRoleRecord(role);
-    }
-  };
-
-  async function deleteSelectedRole(role: ProjectWorkRoleRecord) {
-    const deleted = await onDelete(role);
-    if (!deleted) return;
-    const nextRole = roles.find((item) => !item.built_in && item.id !== role.id) ?? null;
-    if (nextRole) {
-      selectRoleRecord(nextRole);
-      return;
-    }
-    setSelectedRoleID("new");
-    setForm(emptyRoleForm());
-  }
-
-  return (
-    <Modal
-      title="Project roles"
-      onClose={onClose}
-      width={760}
-      footer={
-        <div style={{ display: "flex", gap: 8, width: "100%" }}>
-          {selectedRole && !selectedRole.built_in && !editingNew && (
-            <button
-              className="btn btn-ghost"
-              type="button"
-              disabled={pending}
-              onClick={() => void deleteSelectedRole(selectedRole)}
-              style={{ color: "var(--red)" }}
-            >
-              Delete role
-            </button>
-          )}
-          <button
-            className="btn btn-primary"
-            type="button"
-            disabled={pending || !canSave}
-            onClick={() => void submit()}
-            style={{ marginLeft: "auto" }}
-          >
-            {pending ? "Saving…" : editingNew ? "Create role" : "Save role"}
-          </button>
-        </div>
-      }
-    >
-      <div style={{ display: "grid", gridTemplateColumns: "220px 1fr", gap: 14, minHeight: 420 }}>
-        <div
-          style={{
-            borderRight: "1px solid var(--border)",
-            paddingRight: 10,
-            display: "grid",
-            alignContent: "start",
-            gap: 6,
-          }}
-        >
-          <button
-            className={selectedRoleID === "new" ? "btn btn-primary btn-sm" : "btn btn-ghost btn-sm"}
-            type="button"
-            onClick={() => selectRole("new")}
-            style={{ justifyContent: "flex-start" }}
-          >
-            <Icon d={Icons.plus} size={12} />
-            New custom role
-          </button>
-          {roles.map((role) => (
-            <button
-              key={role.id}
-              className={
-                selectedRoleID === role.id ? "btn btn-primary btn-sm" : "btn btn-ghost btn-sm"
-              }
-              type="button"
-              onClick={() => selectRole(role.id)}
-              style={{ justifyContent: "flex-start", minWidth: 0 }}
-            >
-              <span style={{ overflow: "hidden", textOverflow: "ellipsis" }}>{role.name}</span>
-              {role.built_in && <span className="badge badge-muted">built-in</span>}
-            </button>
-          ))}
-        </div>
-        <form
-          onSubmit={(event) => {
-            event.preventDefault();
-            void submit();
-          }}
-          style={{ display: "grid", gap: 12, alignContent: "start" }}
-        >
-          {error && <InlineError message={error} />}
-          {editingBuiltIn && (
-            <div style={subtleTextStyle}>
-              Built-in roles are read-only. Create a custom role to override instructions or
-              execution defaults for this project.
-            </div>
-          )}
-          <label style={fieldStyle}>
-            <span style={fieldLabelStyle}>Name</span>
-            <input
-              className="input"
-              value={form.name}
-              disabled={editingBuiltIn}
-              onChange={(event) => setForm((current) => ({ ...current, name: event.target.value }))}
-              autoFocus={editingNew}
-            />
-          </label>
-          <label style={fieldStyle}>
-            <span style={fieldLabelStyle}>Description</span>
-            <textarea
-              className="input"
-              value={form.description}
-              disabled={editingBuiltIn}
-              rows={2}
-              onChange={(event) =>
-                setForm((current) => ({ ...current, description: event.target.value }))
-              }
-            />
-          </label>
-          <label style={fieldStyle}>
-            <span style={fieldLabelStyle}>Instructions</span>
-            <textarea
-              className="input"
-              value={form.instructions}
-              disabled={editingBuiltIn}
-              rows={5}
-              onChange={(event) =>
-                setForm((current) => ({ ...current, instructions: event.target.value }))
-              }
-            />
-          </label>
-          <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 10 }}>
-            <label style={fieldStyle}>
-              <span style={fieldLabelStyle}>Default driver</span>
-              <select
-                className="input"
-                value={form.defaultDriverKind}
-                disabled={editingBuiltIn}
-                onChange={(event) =>
-                  setForm((current) => ({ ...current, defaultDriverKind: event.target.value }))
-                }
-              >
-                <option value="">assignment default</option>
-                <option value="hecate_task">hecate_task</option>
-                <option value="external_agent">external_agent</option>
-              </select>
-            </label>
-            <label style={fieldStyle}>
-              <span style={fieldLabelStyle}>Default profile</span>
-              <select
-                className="input"
-                value={form.defaultAgentProfile}
-                disabled={editingBuiltIn}
-                onChange={(event) =>
-                  setForm((current) => ({
-                    ...current,
-                    defaultAgentProfile: event.target.value,
-                  }))
-                }
-              >
-                <option value="">inherit project default</option>
-                {agentProfiles.map((profile) => (
-                  <option key={profile.id} value={profile.id}>
-                    {profile.name || profile.id} ({profile.id})
-                  </option>
-                ))}
-              </select>
-            </label>
-            <label style={fieldStyle}>
-              <span style={fieldLabelStyle}>Default provider</span>
-              <input
-                className="input"
-                value={form.defaultProvider}
-                disabled={editingBuiltIn}
-                placeholder="ollama"
-                onChange={(event) =>
-                  setForm((current) => ({ ...current, defaultProvider: event.target.value }))
-                }
-              />
-            </label>
-            <label style={fieldStyle}>
-              <span style={fieldLabelStyle}>Default model</span>
-              <input
-                className="input"
-                value={form.defaultModel}
-                disabled={editingBuiltIn}
-                placeholder="ministral-3:latest"
-                onChange={(event) =>
-                  setForm((current) => ({ ...current, defaultModel: event.target.value }))
-                }
-              />
-            </label>
-          </div>
-          <SkillIDPicker
-            disabled={editingBuiltIn}
-            onChange={(skillIDs) => setForm((current) => ({ ...current, skillIDs }))}
-            skills={projectSkills}
-            value={form.skillIDs}
-          />
-          <div style={subtleTextStyle}>
-            Role defaults are execution hints. Assignments can still override the driver, and
-            project defaults remain the fallback.
-          </div>
-        </form>
-      </div>
-    </Modal>
-  );
-}
-
 function NewWorkItemModal({
   error,
   pending,
@@ -6301,137 +5547,6 @@ function summarizeAssignments(assignments: ProjectAssignmentRecord[]): WorkItemS
   );
 }
 
-function emptyRoleForm(): RoleForm {
-  return {
-    id: "",
-    name: "",
-    description: "",
-    instructions: "",
-    defaultDriverKind: "",
-    defaultProvider: "",
-    defaultModel: "",
-    defaultAgentProfile: "",
-    skillIDs: "",
-  };
-}
-
-function emptyAgentProfileForm(): AgentProfileForm {
-  return {
-    id: "",
-    name: "",
-    description: "",
-    instructions: "",
-    surface: "any",
-    providerHint: "",
-    modelHint: "",
-    executionProfile: "",
-    toolsEnabled: true,
-    writesAllowed: false,
-    networkAllowed: false,
-    approvalPolicy: "inherit",
-    projectMemoryPolicy: "inherit",
-    contextSourcePolicy: "inherit",
-    skillIDs: "",
-    externalAgentKind: "",
-  };
-}
-
-function profileFormFromRecord(profile: AgentProfileRecord): AgentProfileForm {
-  return {
-    id: profile.id,
-    name: profile.name,
-    description: profile.description ?? "",
-    instructions: profile.instructions ?? "",
-    surface: profile.surface || "any",
-    providerHint: profile.provider_hint ?? "",
-    modelHint: profile.model_hint ?? "",
-    executionProfile: profile.execution_profile ?? "",
-    toolsEnabled: profile.tools_enabled,
-    writesAllowed: profile.writes_allowed,
-    networkAllowed: profile.network_allowed,
-    approvalPolicy: profile.approval_policy || "inherit",
-    projectMemoryPolicy: profile.project_memory_policy || "inherit",
-    contextSourcePolicy: profile.context_source_policy || "inherit",
-    skillIDs: (profile.skill_ids ?? []).join(", "),
-    externalAgentKind: profile.external_agent_kind ?? "",
-  };
-}
-
-function profileCreatePayloadFromForm(form: AgentProfileForm): CreateAgentProfilePayload {
-  const payload = profileUpdatePayloadFromForm(form) as CreateAgentProfilePayload;
-  const id = form.id.trim();
-  if (id) payload.id = id;
-  return payload;
-}
-
-function profileUpdatePayloadFromForm(form: AgentProfileForm): UpdateAgentProfilePayload {
-  return {
-    name: form.name.trim(),
-    description: form.description.trim(),
-    instructions: form.instructions.trim(),
-    surface: form.surface.trim() || "any",
-    provider_hint: form.providerHint.trim(),
-    model_hint: form.modelHint.trim(),
-    execution_profile: form.executionProfile.trim(),
-    tools_enabled: form.toolsEnabled,
-    writes_allowed: form.writesAllowed,
-    network_allowed: form.networkAllowed,
-    approval_policy: form.approvalPolicy.trim() || "inherit",
-    project_memory_policy: form.projectMemoryPolicy.trim() || "inherit",
-    context_source_policy: form.contextSourcePolicy.trim() || "inherit",
-    skill_ids: uniqueSkillIDs(splitIDs(form.skillIDs)),
-    external_agent_kind: form.externalAgentKind.trim(),
-  };
-}
-
-function profileReferenceSummary(
-  profile: AgentProfileRecord,
-  project: ProjectRecord,
-  roles: ProjectWorkRoleRecord[],
-) {
-  const references: string[] = [];
-  if (project.default_agent_profile === profile.id) {
-    references.push("this project's default profile");
-  }
-  const roleNames = roles
-    .filter((role) => role.default_agent_profile === profile.id)
-    .map((role) => role.name || role.id);
-  if (roleNames.length > 0) {
-    references.push(`roles ${roleNames.join(", ")}`);
-  }
-  if (references.length === 0) {
-    return "No current project defaults or roles reference it.";
-  }
-  return `Referenced by ${references.join("; ")}. Those references will fall back until changed.`;
-}
-
-function roleFormFromRecord(role: ProjectWorkRoleRecord): RoleForm {
-  return {
-    id: role.id,
-    name: role.name,
-    description: role.description ?? "",
-    instructions: role.instructions ?? "",
-    defaultDriverKind: role.default_driver_kind ?? "",
-    defaultProvider: role.default_provider ?? "",
-    defaultModel: role.default_model ?? "",
-    defaultAgentProfile: role.default_agent_profile ?? "",
-    skillIDs: (role.skill_ids ?? []).join(", "),
-  };
-}
-
-function rolePayloadFromForm(form: RoleForm) {
-  return {
-    name: form.name.trim(),
-    description: form.description.trim(),
-    instructions: form.instructions.trim(),
-    default_driver_kind: form.defaultDriverKind.trim(),
-    default_provider: form.defaultProvider.trim(),
-    default_model: form.defaultModel.trim(),
-    default_agent_profile: form.defaultAgentProfile.trim(),
-    skill_ids: uniqueSkillIDs(splitIDs(form.skillIDs)),
-  };
-}
-
 function skillFormFromRecord(skill: ProjectSkillRecord): SkillForm {
   return {
     title: skill.title ?? "",
@@ -6792,52 +5907,6 @@ function upsertProjectSkill(items: ProjectSkillRecord[], item: ProjectSkillRecor
   });
 }
 
-function sortProjectSkillsForPicker(skills: ProjectSkillRecord[]) {
-  return skills.slice().sort((a, b) => {
-    if (a.enabled !== b.enabled) return a.enabled ? -1 : 1;
-    const rank = projectSkillStatusRank(a.status) - projectSkillStatusRank(b.status);
-    return (
-      rank ||
-      a.title.localeCompare(b.title) ||
-      a.path.localeCompare(b.path) ||
-      a.id.localeCompare(b.id)
-    );
-  });
-}
-
-function projectSkillBadgeClass(skill: ProjectSkillRecord) {
-  if (skill.status === "available" && skill.enabled) return "badge badge-green";
-  if (skill.status === "available") return "badge badge-muted";
-  return "badge badge-amber";
-}
-
-function projectSkillSelectionWarnings(
-  skillID: string,
-  indexedSkills: Map<string, ProjectSkillRecord>,
-) {
-  const skill = indexedSkills.get(skillID);
-  if (!skill) return [`Skill ${skillID} is not registered in this project.`];
-  const warnings: string[] = [];
-  if (!skill.enabled) warnings.push(`Skill ${skillID} is disabled.`);
-  if (skill.status !== "available") warnings.push(`Skill ${skillID} is ${skill.status}.`);
-  return warnings;
-}
-
-function projectSkillStatusRank(status: string): number {
-  switch (status) {
-    case "available":
-      return 0;
-    case "conflict":
-      return 1;
-    case "invalid":
-      return 2;
-    case "missing":
-      return 3;
-    default:
-      return 4;
-  }
-}
-
 function upsertWorkItem(items: ProjectWorkItemRecord[], item: ProjectWorkItemRecord) {
   const index = items.findIndex((current) => current.id === item.id);
   if (index === -1) return [item, ...items];
@@ -6907,17 +5976,6 @@ function rememberRightPanelWidth(width: number) {
 
 function splitRoleIDs(value: string): string[] {
   return splitIDs(value);
-}
-
-function splitIDs(value: string): string[] {
-  return value
-    .split(",")
-    .map((item) => item.trim())
-    .filter(Boolean);
-}
-
-function uniqueSkillIDs(ids: string[]): string[] {
-  return Array.from(new Set(ids));
 }
 
 function errorMessage(error: unknown, fallback: string): string {
@@ -7063,14 +6121,6 @@ const fieldLabelStyle: CSSProperties = {
   fontFamily: "var(--font-mono)",
   fontSize: 11,
   textTransform: "uppercase",
-};
-
-const checkboxLabelStyle: CSSProperties = {
-  alignItems: "center",
-  color: "var(--t1)",
-  display: "inline-flex",
-  fontSize: 12,
-  gap: 6,
 };
 
 const panelStyle: CSSProperties = {
